+++ tests/sdk/tool/test_tool_serialization.py
@@ -1,182 +1,149 @@
-"""Test tool serialization with DiscriminatedUnionMixin."""
+"""Test tool JSON serialization with DiscriminatedUnionMixin."""
 
-from typing import Annotated
+import json
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel
 
 from openhands.sdk.tool import Tool, ToolType
 from openhands.sdk.tool.builtins import FinishTool, ThinkTool
-from openhands.sdk.tool.schema import ActionBase, ObservationBase
-from openhands.sdk.utils.discriminated_union import DiscriminatedUnionType
-from openhands.tools.execute_bash.definition import BashTool
 
 
-def test_tool_supports_polymorphic_deserialization() -> None:
-    """Test that Tool supports polymorphic deserialization from dict data."""
-    # Use FinishTool as a concrete Tool instance
+def test_tool_serialization_deserialization() -> None:
+    """Test that Tool supports polymorphic JSON serialization/deserialization."""
+    # Use FinishTool which is a simple built-in tool
     tool = FinishTool
 
-    # Get the serialized data as dict (not JSON due to non-serializable fields)
-    tool_data = tool.model_dump()
+    # Serialize to JSON
+    tool_json = tool.model_dump_json()
 
-    # Deserialize using the base class
-    deserialized_tool = Tool.model_validate(tool_data)
+    # Deserialize from JSON using the base class
+    deserialized_tool = Tool.model_validate_json(tool_json)
 
-    # Should deserialize to the correct concrete type
-    assert isinstance(deserialized_tool, type(tool))
-    assert deserialized_tool.name == tool.name
-    assert deserialized_tool.description == tool.description
-    assert deserialized_tool.action_type == tool.action_type
-    assert deserialized_tool.observation_type == tool.observation_type
+    # Should deserialize to the correct type with same serializable data
+    assert isinstance(deserialized_tool, Tool)
+    assert tool.model_dump() == deserialized_tool.model_dump()
 
 
-def test_tool_supports_polymorphic_field_deserialization() -> None:
-    """Test that Tool supports polymorphic deserialization when used as a field."""
+def test_tool_supports_polymorphic_field_json_serialization() -> None:
+    """Test that Tool supports polymorphic JSON serialization when used as a field."""
 
-    class ToolContainer(BaseModel):
-        tool: Annotated[Tool, DiscriminatedUnionType[Tool]]
+    class Container(BaseModel):
+        tool: Tool
 
-    # Use ThinkTool as a concrete Tool instance
-    tool = ThinkTool
-    container = ToolContainer(tool=tool)
-
-    # Get dict data (not JSON due to non-serializable fields)
-    container_data = container.model_dump()
-    deserialized_container = ToolContainer.model_validate(container_data)
-
-    # Should deserialize to the correct concrete type
-    assert isinstance(deserialized_container.tool, type(tool))
-    assert deserialized_container.tool.name == tool.name
-    assert deserialized_container.tool.description == tool.description
-
-
-def test_tool_supports_nested_polymorphic_deserialization() -> None:
-    """Test that Tool supports polymorphic deserialization when nested in lists."""
-
-    class ToolRegistry(BaseModel):
-        tools: list[Annotated[Tool, DiscriminatedUnionType[Tool]]]
-
-    # Use built-in tools as concrete Tool instances
-    tools = [FinishTool, ThinkTool]
-    registry = ToolRegistry(tools=tools)
-
-    # Get dict data (not JSON due to non-serializable fields)
-    registry_data = registry.model_dump()
-    deserialized_registry = ToolRegistry.model_validate(registry_data)
-
-    # Should deserialize to the correct concrete types
-    assert len(deserialized_registry.tools) == 2
-    assert isinstance(deserialized_registry.tools[0], type(FinishTool))
-    assert isinstance(deserialized_registry.tools[1], type(ThinkTool))
-    assert deserialized_registry.tools[0].name == "finish"
-    assert deserialized_registry.tools[1].name == "think"
-
-
-def test_tool_model_validate_dict() -> None:
-    """Test Tool model_validate with dictionary input."""
-    # Use FinishTool as a concrete Tool instance
+    # Create container with tool
     tool = FinishTool
-    tool_data = tool.model_dump()
+    container = Container(tool=tool)
 
-    # Test with valid kind
-    result = Tool.model_validate(tool_data)
-    assert isinstance(result, type(tool))
-    assert result.name == tool.name
-    assert result.description == tool.description
-    assert result.action_type == tool.action_type
-    assert result.observation_type == tool.observation_type
+    # Serialize to JSON
+    container_json = container.model_dump_json()
+
+    # Deserialize from JSON
+    deserialized_container = Container.model_validate_json(container_json)
+
+    # Should preserve the tool type with same serializable data
+    assert isinstance(deserialized_container.tool, Tool)
+    assert tool.model_dump() == deserialized_container.tool.model_dump()
 
 
-def test_tool_fallback_behavior() -> None:
-    """Test fallback behavior when discriminated union logic doesn't apply."""
+def test_tool_supports_nested_polymorphic_json_serialization() -> None:
+    """Test that Tool supports nested polymorphic JSON serialization."""
 
-    # Create a minimal tool data without kind
-    class TestAction(ActionBase):
-        test_field: str = Field(description="Test field")
+    class NestedContainer(BaseModel):
+        tools: list[Tool]
 
-    class TestObservation(ObservationBase):
-        result: str = Field(description="Test result")
+    # Create container with multiple tools
+    tool1 = FinishTool
+    tool2 = ThinkTool
+    container = NestedContainer(tools=[tool1, tool2])
 
-    no_kind_data = {
-        "name": "test_tool",
+    # Serialize to JSON
+    container_json = container.model_dump_json()
+
+    # Deserialize from JSON
+    deserialized_container = NestedContainer.model_validate_json(container_json)
+
+    # Should preserve all tool types with same serializable data
+    assert len(deserialized_container.tools) == 2
+    assert isinstance(deserialized_container.tools[0], Tool)
+    assert isinstance(deserialized_container.tools[1], Tool)
+    assert tool1.model_dump() == deserialized_container.tools[0].model_dump()
+    assert tool2.model_dump() == deserialized_container.tools[1].model_dump()
+
+
+def test_tool_model_validate_json_dict() -> None:
+    """Test that Tool.model_validate works with dict from JSON."""
+    # Create tool
+    tool = FinishTool
+
+    # Serialize to JSON, then parse to dict
+    tool_json = tool.model_dump_json()
+    tool_dict = json.loads(tool_json)
+
+    # Deserialize from dict
+    deserialized_tool = Tool.model_validate(tool_dict)
+
+    # Should have same serializable data
+    assert isinstance(deserialized_tool, Tool)
+    assert tool.model_dump() == deserialized_tool.model_dump()
+
+
+def test_tool_fallback_behavior_json() -> None:
+    """Test that Tool handles unknown types gracefully in JSON."""
+    # Create JSON with unknown kind
+    tool_dict = {
+        "name": "test-tool",
         "description": "A test tool",
-        "action_type": TestAction,
-        "observation_type": TestObservation,
+        "action_type": "openhands.sdk.tool.builtins.finish.FinishAction",
+        "observation_type": None,
+        "kind": "UnknownToolType",
     }
+    tool_json = json.dumps(tool_dict)
 
-    # Test with missing kind - should fallback to base class
-    result = Tool.model_validate(no_kind_data)
-    assert isinstance(result, Tool)
-    assert result.name == "test_tool"
-
-    # Test with invalid kind - should fallback to base class
-    invalid_kind_data = {**no_kind_data, "kind": "InvalidTool"}
-    result = Tool.model_validate(invalid_kind_data)
-    assert isinstance(result, Tool)
-    assert result.name == "test_tool"
+    # Should fall back to base Tool type
+    deserialized_tool = Tool.model_validate_json(tool_json)
+    assert isinstance(deserialized_tool, Tool)
+    assert deserialized_tool.name == "test-tool"
+    assert deserialized_tool.description == "A test tool"
 
 
-def test_tool_preserves_pydantic_parameters() -> None:
-    """Test that all Pydantic validation parameters are preserved."""
+def test_tool_type_annotation_works_json() -> None:
+    """Test that ToolType annotation works correctly with JSON."""
+    # Create tool
     tool = FinishTool
-    tool_data = tool.model_dump()
 
-    # Test with strict mode
-    result = Tool.model_validate(tool_data, strict=True)
-    assert isinstance(result, type(tool))
-
-    # Test with context (even though we don't use it here, it should be passed through)
-    context = {"test": "value"}
-    result = Tool.model_validate(tool_data, context=context)
-    assert isinstance(result, type(tool))
-
-
-def test_tool_type_annotation_works() -> None:
-    """Test that ToolType annotation works correctly."""
-
-    class ToolWrapper(BaseModel):
+    # Use ToolType annotation
+    class TestModel(BaseModel):
         tool: ToolType
 
-    tool = ThinkTool
-    wrapper = ToolWrapper(tool=tool)
+    model = TestModel(tool=tool)
 
-    # Get dict data (not JSON due to non-serializable fields)
-    wrapper_data = wrapper.model_dump()
-    deserialized_wrapper = ToolWrapper.model_validate(wrapper_data)
+    # Serialize to JSON
+    model_json = model.model_dump_json()
 
-    # Should deserialize to the correct concrete type
-    assert isinstance(deserialized_wrapper.tool, type(tool))
-    assert deserialized_wrapper.tool.name == tool.name
+    # Deserialize from JSON
+    deserialized_model = TestModel.model_validate_json(model_json)
 
-
-def test_custom_tool_subclass_deserialization() -> None:
-    """Test that custom Tool subclasses deserialize correctly."""
-
-    # Create a BashTool instance (which is a Tool subclass)
-    bash_tool = BashTool.create(working_dir="/tmp")
-
-    # Get dict data (not JSON due to non-serializable fields)
-    tool_data = bash_tool.model_dump()
-    deserialized_tool = Tool.model_validate(tool_data)
-
-    # Should deserialize to the correct concrete type
-    assert isinstance(deserialized_tool, BashTool)
-    assert deserialized_tool.name == bash_tool.name
-    assert deserialized_tool.description == bash_tool.description
-    assert deserialized_tool.action_type == bash_tool.action_type
-    assert deserialized_tool.observation_type == bash_tool.observation_type
+    # Should work correctly with same serializable data
+    assert isinstance(deserialized_model.tool, Tool)
+    assert tool.model_dump() == deserialized_model.tool.model_dump()
 
 
-def test_tool_with_annotations_deserialization() -> None:
-    """Test that Tool with annotations deserializes correctly."""
-    # FinishTool has annotations
+def test_tool_kind_field_json() -> None:
+    """Test Tool kind field is correctly set and preserved through JSON."""
+    # Create tool
     tool = FinishTool
 
-    # Get dict data (not JSON due to non-serializable fields)
-    tool_data = tool.model_dump()
-    deserialized_tool = Tool.model_validate(tool_data)
+    # Check kind field
+    assert hasattr(tool, "kind")
+    expected_kind = f"{tool.__class__.__module__}.{tool.__class__.__name__}"
+    assert tool.kind == expected_kind
 
-    # Should preserve annotations
-    assert deserialized_tool.annotations == tool.annotations
-    # Basic check that annotations exist
-    assert deserialized_tool.annotations is not None
+    # Serialize to JSON
+    tool_json = tool.model_dump_json()
+
+    # Deserialize from JSON
+    deserialized_tool = Tool.model_validate_json(tool_json)
+
+    # Should preserve kind field
+    assert hasattr(deserialized_tool, "kind")
+    assert deserialized_tool.kind == tool.kind
